Исправления замечаний
